add call stats provider
