+++ lib/main.dart
@@ -1,9 +1,11 @@
+import 'package:callstats/providers/call_stats_provider.dart';
 import 'package:callstats/screens/single_person_call_stats_screen.dart';
 import 'package:callstats/screens/home_screen.dart';
 import 'package:callstats/screens/landing_screen.dart';
 import 'package:callstats/screens/splash_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:provider/provider.dart';
 
 void main() {
   runApp(const MyApp());
@@ -25,29 +27,32 @@ class _MyAppState extends State<MyApp> {
         statusBarIconBrightness: Brightness.dark,
       ),
     );
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      initialRoute: SplashScreen.routeName,
-      routes: {
-        SplashScreen.routeName: (context) => const SplashScreen(),
-        LandingScreen.routeName: (context) => const LandingScreen(),
-        HomeScreen.routeName: (context) => const HomeScreen(),
-        SinglePersonCallStatsScreen.routeName: (context) =>
-            const SinglePersonCallStatsScreen(),
-      },
-      theme: ThemeData(
-        scaffoldBackgroundColor: const Color(0xFFEEEEEE),
-        appBarTheme: AppBarTheme(
-          backgroundColor: Colors.grey[200],
-          elevation: 0.0,
-          titleTextStyle: const TextStyle(
-            color: Colors.black,
-            fontSize: 20.0,
+    return ChangeNotifierProvider(
+      create: (_) => CallStatsProvider(),
+      child: MaterialApp(
+        debugShowCheckedModeBanner: false,
+        initialRoute: SplashScreen.routeName,
+        routes: {
+          SplashScreen.routeName: (context) => const SplashScreen(),
+          LandingScreen.routeName: (context) => const LandingScreen(),
+          HomeScreen.routeName: (context) => const HomeScreen(),
+          SinglePersonCallStatsScreen.routeName: (context) =>
+              const SinglePersonCallStatsScreen(),
+        },
+        theme: ThemeData(
+          scaffoldBackgroundColor: const Color(0xFFEEEEEE),
+          appBarTheme: AppBarTheme(
+            backgroundColor: Colors.grey[200],
+            elevation: 0.0,
+            titleTextStyle: const TextStyle(
+              color: Colors.black,
+              fontSize: 20.0,
+            ),
+            actionsIconTheme: const IconThemeData(
+              color: Colors.black,
+            ),
+            foregroundColor: Colors.black,
           ),
-          actionsIconTheme: const IconThemeData(
-            color: Colors.black,
-          ),
-          foregroundColor: Colors.black,
         ),
       ),
     );
